@@ -4,46 +4,64 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build",
+    "build": "next build ",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "test": "jest --watch"
   },
   "dependencies": {
     "@cfdi/xml": "^4.0.13",
     "@fortawesome/fontawesome-free": "^6.5.1",
-    "@signati/saxon": "^0.0.4",
     "@types/bcryptjs": "^2.4.6",
+    "@types/mongoose": "^5.11.96",
+    "aws4": "^1.13.2",
     "axios": "^1.6.7",
     "bcryptjs": "^2.4.3",
     "bootstrap": "^5.3.3",
     "cfdi40": "^3.3.6",
+    "exceljs": "^4.4.0",
+    "express": "^4.21.2",
     "jstoxml": "^5.0.2",
     "mdb-react-ui-kit": "^7.2.0",
     "mongoose": "^8.5.4",
     "next": "14.1.0",
-    "next-auth": "^4.24.7",
+    "next-auth": "^4.24.11",
     "order-id": "^2.1.2",
-    "postgres": "^3.4.3",
+    "pdfkit": "^0.16.0",
+    "pdfkit-table": "^0.1.99",
     "react": "^18",
-    "react-bootstrap": "^2.10.2",
+    "react-bootstrap": "^2.10.7",
     "react-dom": "^18",
     "react-hook-form": "^7.50.1",
+    "react-hot-toast": "^2.5.1",
+    "react-icons": "^5.4.0",
     "react-select": "^5.8.0",
     "react-select-search": "^4.1.7",
-    "receipt": "^1.4.0",
+    "react-toastify": "^11.0.2",
     "sequelize": "^6.37.5",
     "sqlite3": "^5.1.7",
     "sw-sdk-nodejs": "^1.1.4"
   },
   "devDependencies": {
-    "@types/node": "^20",
+    "@babel/core": "^7.26.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^14.0.0",
+    "@types/express": "^5.0.0",
+    "@types/jest": "^29.0.0",
+    "@types/node": "^20.17.12",
+    "@types/pdfkit": "^0.13.8",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@types/testing-library__jest-dom": "^5.14.9",
     "autoprefixer": "^10.0.1",
+    "babel-jest": "^29.7.0",
+    "browserslist": "^4.24.4",
     "eslint": "^8",
     "eslint-config-next": "14.1.0",
+    "jest": "^29.0.0",
+    "jest-environment-jsdom": "^29.7.0",
     "postcss": "^8",
     "tailwindcss": "^3.3.0",
-    "typescript": "^5"
+    "typescript": "^5.7.3"
   }
 }
