@@ -1,10 +1,15 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
+import "./styles/fixes.css";
+import 'bootstrap/dist/css/bootstrap.min.css'
+import { ToastContainer } from 'react-toastify'
+import 'react-toastify/dist/ReactToastify.css'
+import { NextAuthProvider } from "./providers";
 
 const inter = Inter({ subsets: ["latin"] });
 export const metadata: Metadata = {
-  title: "Bici-Vic",
+  title: "Rally ERP",
   description: "Generated by create next app",
 };
 import NavScroll from "./uiComponents/nav";
@@ -12,8 +17,12 @@ import NavScroll from "./uiComponents/nav";
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en">
-      <body>
-        <main>{children}</main>
+      <body className={inter.className}>
+        <NextAuthProvider>
+          <NavScroll />
+          <main>{children}</main>
+        </NextAuthProvider>
+        <ToastContainer position="top-right" autoClose={3000} />
       </body>
     </html>
   )
