@@ -1,177 +1,360 @@
-'use server'
-import postgres from 'postgres'
-import { useForm } from "react-hook-form";
-import axios from "axios";
-import 'bootstrap/dist/css/bootstrap.min.css';
-import Lista from "../../uiComponents/lista";
-import FormCreator from "../../uiComponents/formCreator";
- let url =axios.defaults.baseURL = 'https://express-low5.onrender.com'
-/*const sql = postgres( 'postgres://robbie:1234@localhost:5432/mydb ') // will use psql environment variables
-    async function getNotes() {
-      let notes = await sql`SELECT id,partner_id, fiscal_position_id, payment_term_id, pricelist_id, create_uid, write_uid, name, state, client_order_ref, origin, reference, signed_by, invoice_status, validity_date, note, currency_rate, amount_untaxed, amount_tax, amount_total, amount_to_invoice, locked, require_signature, require_payment, create_date, commitment_date, date_order, signed_on, write_date, prepayment_percent, sale_order_template_id
-      FROM public.sale_order WHERE partner_id = '44';
-      `
-      // notes = Result [{ name: "Walter", age: 80 }, { name: 'Murray', age: 68 }, ...]
-      return notes
+'use client'
+import { useState, useEffect } from 'react'
+import { useRouter } from 'next/navigation'
+
+interface Direccion {
+  calle: string;
+  numeroExterior: string;
+  numeroInterior: string;
+  colonia: string;
+  municipio: string;
+  estado: string;
+}
+
+interface FormData {
+  razonSocial: string;
+  rfc: string;
+  regimenFiscal: string;
+  codigoPostal: string;
+  registroPatronal: string;
+  curp: string;
+  direccion: Direccion;
+  telefono: string;
+  correoElectronico: string;
+}
+
+export default function DatosNegocio() {
+  const router = useRouter()
+  const [loading, setLoading] = useState(true)
+  const [saving, setSaving] = useState(false)
+  const [error, setError] = useState('')
+  const [formData, setFormData] = useState<FormData>({
+    razonSocial: '',
+    rfc: '',
+    regimenFiscal: '',
+    codigoPostal: '',
+    registroPatronal: '',
+    curp: '',
+    direccion: {
+      calle: '',
+      numeroExterior: '',
+      numeroInterior: '',
+      colonia: '',
+      municipio: '',
+      estado: ''
+    },
+    telefono: '',
+    correoElectronico: ''
+  })
+
+  useEffect(() => {
+    fetchDatosNegocio()
+  }, [])
+
+  const fetchDatosNegocio = async () => {
+    try {
+      const response = await fetch('/api/datosnegocio')
+      if (response.ok) {
+        const data = await response.json()
+        if (data) {
+          setFormData({
+            ...data,
+            direccion: {
+              calle: data.direccion?.calle || '',
+              numeroExterior: data.direccion?.numeroExterior || '',
+              numeroInterior: data.direccion?.numeroInterior || '',
+              colonia: data.direccion?.colonia || '',
+              municipio: data.direccion?.municipio || '',
+              estado: data.direccion?.estado || ''
+            }
+          })
+        }
+      }
+    } catch (error) {
+      setError('Error al cargar los datos del negocio')
+    } finally {
+      setLoading(false)
     }
-  
-
-
-
-let notes = getNotes()
-let  notasPartidas: {
-  clave_producto_servicio: string; clave_unidad_de_medida: string; cantidad: number; descripcion: string; valor_unitario: number; total // will use psql environment variables
-    : number; exento_de_impuestos: boolean; impuestos: { traslados: { traslado: { base: any; impuesto: string; tipoFactor: string; tasaOCuota: number; importe: number; tipo: string; }[]; }; };
-}[]=  [];
-
-function addPartida(notaManual: any) {
-     
-  const importe =  notaManual.amount_total
-let folio = notaManual.name
-let iva = Number(importe) * 0.16
-
-
-  let partida= {
-   
-    clave_producto_servicio: "01010101",
-    clave_unidad_de_medida: "ACT",
-    cantidad: 1,
-    descripcion: "Venta",
-    valor_unitario: Number(importe),
-    total: Number(importe),
-    exento_de_impuestos: false,
-    impuestos: {
-      traslados: {
-      traslado: [
-      {
-      base: importe,
-      impuesto: "002",
-      tipoFactor: "Tasa",
-      tasaOCuota: 0.16,
-      importe: Number(importe) - iva,
-      tipo: "Traslado"
-      }
-      ]
-      }
-   }
-   
   }
-  notasPartidas.push(partida)
 
-console.log(notasPartidas);
-}
+  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const { name, value } = e.target
+    if (name.includes('.')) {
+      const [parent, child] = name.split('.')
+      if (parent === 'direccion') {
+        setFormData(prev => ({
+          ...prev,
+          direccion: {
+            ...prev.direccion,
+            [child]: value
+          }
+        }))
+      }
+    } else {
+      setFormData(prev => ({
+        ...prev,
+        [name]: value
+      }))
+    }
+  }
 
-notes.then((notes) => notes.map((note) => {
-  let sec =  Date.now();
-  let hoy = new Date(sec).toISOString();
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault()
+    setSaving(true)
+    setError('')
 
-  let fechaCreacion = new Date(note.create_date).toISOString().split('T')[0];
-  
-  
-console.log(hoy,note.name,fechaCreacion,
-  note.amount_untaxed,note.amount_total,note.invoice_status,note.state); 
-  
-   namecheap pone la fecha de creacion 12 horas despues
-  console.log(note);
-  
-  if (note.partner_id == '44' && fechaCreacion== hoy && note.invoice_status == 'to invoice') {
+    try {
+      const response = await fetch('/api/datosnegocio', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(formData),
+      })
 
+      if (!response.ok) {
+        const data = await response.json()
+        throw new Error(data.error || 'Error al guardar los datos')
+      }
 
-    addPartida(note);
-    console.log(notasPartidas);
-}
+      router.refresh()
+    } catch (error: any) {
+      setError(error.message || 'Error al guardar los datos del negocio')
+    } finally {
+      setSaving(false)
+    }
+  }
 
-
-}
-
-)
-
-)
-
-*/
-let ListaProps= {
-  // Define the properties and their types here
-  'precio' : String,
-  'categoria' : String,
-  'descripcion' : String,
-  'marca' : String,
-  'especificaciones' : String,
-  'codigo' : String,
-  'moto' : String,
-  'compatibilidades' : [],
-  'recurso' : String
-}
-
-let elements= [{
-  name: "precio",
-  id: 1,
-  type: "text",
-  placeholder: "precio"
-
-} ,
-{
-  name: "categoria",
-  id: 2,
-  type: "text",
-  placeholder: "categoria"
-},
-{
-  name: "descripcion",
-  id: 3,
-  type: "text",
-  placeholder: "descripcion"
-},
-
-{
-  name: "marca",
-  id: 4,
-  type: "text",
-  placeholder: "marca"
-},
-{
-  name: "especificaciones",
-  id: 5,
-  type: "text",
-  placeholder: "especificaciones"
-},
-{
-  name: "codigo",
-  id: 6,
-  type: "text",
-  placeholder: "codigo"
-},
-{
-  name: "moto",
-  id: 7,
-  type: "text",
-  placeholder: "moto"
-},
-
-{
-  name: "compatibilidades",
-  id: 8,
-  type: "text",
-  placeholder: "compatibilidades"
-}
-]
-export default async function Productos() {
-  // Render the main container
-  return (
-  
-      <div>
-        
-        {/* Render the form creator */}
-        <FormCreator 
-          // Pass the elements array as a prop
-          elements={elements} 
-          // Pass the URL for the productos resource as a prop
-          recurso={url+"/productos"} 
-          image={true}
-        />
-        {/* Render the list component */}
-        
-        <Lista recurso='/productos'  />
+  if (loading) {
+    return (
+      <div className="flex justify-center items-center min-h-screen">
+        <div className="text-center">Cargando...</div>
       </div>
-    
-  );
+    )
+  }
+
+  return (
+    <div className="container mx-auto px-4 py-8">
+      <h1 className="text-2xl font-bold mb-6">Datos del Negocio</h1>
+      
+      {error && (
+        <div className="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded mb-4">
+          {error}
+        </div>
+      )}
+
+      <form onSubmit={handleSubmit} className="space-y-6">
+        <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+          <div>
+            <label className="block text-sm font-medium text-gray-700">
+              Razón Social
+            </label>
+            <input
+              type="text"
+              name="razonSocial"
+              value={formData.razonSocial}
+              onChange={handleInputChange}
+              className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
+              required
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700">
+              RFC
+            </label>
+            <input
+              type="text"
+              name="rfc"
+              value={formData.rfc}
+              onChange={handleInputChange}
+              className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
+              required
+              minLength={12}
+              maxLength={13}
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700">
+              Régimen Fiscal
+            </label>
+            <input
+              type="text"
+              name="regimenFiscal"
+              value={formData.regimenFiscal}
+              onChange={handleInputChange}
+              className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
+              required
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700">
+              Código Postal
+            </label>
+            <input
+              type="text"
+              name="codigoPostal"
+              value={formData.codigoPostal}
+              onChange={handleInputChange}
+              className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
+              required
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700">
+              Registro Patronal
+            </label>
+            <input
+              type="text"
+              name="registroPatronal"
+              value={formData.registroPatronal}
+              onChange={handleInputChange}
+              className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
+              required
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700">
+              CURP
+            </label>
+            <input
+              type="text"
+              name="curp"
+              value={formData.curp}
+              onChange={handleInputChange}
+              className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
+              required
+              minLength={18}
+              maxLength={18}
+            />
+          </div>
+        </div>
+
+        <div className="bg-gray-50 p-4 rounded-md">
+          <h2 className="text-lg font-medium mb-4">Dirección</h2>
+          <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+            <div>
+              <label className="block text-sm font-medium text-gray-700">
+                Calle
+              </label>
+              <input
+                type="text"
+                name="direccion.calle"
+                value={formData.direccion.calle}
+                onChange={handleInputChange}
+                className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700">
+                Número Exterior
+              </label>
+              <input
+                type="text"
+                name="direccion.numeroExterior"
+                value={formData.direccion.numeroExterior}
+                onChange={handleInputChange}
+                className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700">
+                Número Interior
+              </label>
+              <input
+                type="text"
+                name="direccion.numeroInterior"
+                value={formData.direccion.numeroInterior}
+                onChange={handleInputChange}
+                className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700">
+                Colonia
+              </label>
+              <input
+                type="text"
+                name="direccion.colonia"
+                value={formData.direccion.colonia}
+                onChange={handleInputChange}
+                className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700">
+                Municipio
+              </label>
+              <input
+                type="text"
+                name="direccion.municipio"
+                value={formData.direccion.municipio}
+                onChange={handleInputChange}
+                className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
+              />
+            </div>
+
+            <div>
+              <label className="block text-sm font-medium text-gray-700">
+                Estado
+              </label>
+              <input
+                type="text"
+                name="direccion.estado"
+                value={formData.direccion.estado}
+                onChange={handleInputChange}
+                className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
+              />
+            </div>
+          </div>
+        </div>
+
+        <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
+          <div>
+            <label className="block text-sm font-medium text-gray-700">
+              Teléfono
+            </label>
+            <input
+              type="tel"
+              name="telefono"
+              value={formData.telefono}
+              onChange={handleInputChange}
+              className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700">
+              Correo Electrónico
+            </label>
+            <input
+              type="email"
+              name="correoElectronico"
+              value={formData.correoElectronico}
+              onChange={handleInputChange}
+              className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
+            />
+          </div>
+        </div>
+
+        <div className="flex justify-end">
+          <button
+            type="submit"
+            disabled={saving}
+            className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline disabled:opacity-50"
+          >
+            {saving ? 'Guardando...' : 'Guardar Cambios'}
+          </button>
+        </div>
+      </form>
+    </div>
+  )
 }
