@@ -1,39 +1,34 @@
 'use client'
-import getNotes from "../../libComponents/getNotesFromO";
-import FormFactCliente from "@/app/uiComponents/formFactCliente";
+import FormFactCliente from "@/app/uiComponents/formFactClienteNew";
 import NavScroll from "@/app/uiComponents/nav";
- function Page() {
-    
-    
-    
-      
- let elements =
-  [{
-    name: "Descripcion",
-    id: 1,
-    type: "text",
-    placeholder: "Descripcion"
-  
-  } ,
-  {
-    name: "importe",
-    id: 2,
-    type: "number",
-    placeholder: "Importe"},
-    {
-      name: "claveprodserv",
-      id: 3,
-      type: "text",
-      placeholder: "Clave SAT"}
-  ]
-    return (
-        <div>
-          <NavScroll/>
-         <FormFactCliente elements={elements} />
 
+function Page() {
+  const partidas = [{
+    ClaveProdServ: "",
+    ClaveUnidad: "H87",
+    Unidad: "Pieza",
+    Cantidad: "1",
+    Descripcion: "",
+    ValorUnitario: "0",
+    Importe: "0",
+    ImporteRealConImp: "0",
+    ObjetoImp: "02",
+    Impuestos: {
+      Traslados: [{
+        Base: "0",
+        Impuesto: "002",
+        TipoFactor: "Tasa",
+        TasaOCuota: "0.160000",
+        Importe: "0"
+      }]
+    }
+  }];
+
+  return (
+    <div>
+      <FormFactCliente {...partidas} />
     </div>
-    );
-  
-  }
-  
-  export default Page
+  );
+}
+
+export default Page;
