@@ -1,21 +1,23 @@
 'use server'
-import postgres from 'postgres'
 import { useForm } from "react-hook-form";
 import axios from "axios";
 import 'bootstrap/dist/css/bootstrap.min.css';
 import Lista from "../../uiComponents/lista";
 import FormCreator from "../../uiComponents/formCreator";
-/*const sql = postgres( 'postgres://robbie:1234@localhost:5432/mydb ') // will use psql environment variables
-    async function getNotes() {
-      let notes = await sql`SELECT id,partner_id, fiscal_position_id, payment_term_id, pricelist_id, create_uid, write_uid, name, state, client_order_ref, origin, reference, signed_by, invoice_status, validity_date, note, currency_rate, amount_untaxed, amount_tax, amount_total, amount_to_invoice, locked, require_signature, require_payment, create_date, commitment_date, date_order, signed_on, write_date, prepayment_percent, sale_order_template_id
-      FROM public.sale_order WHERE partner_id = '44';
-      `
-      // notes = Result [{ name: "Walter", age: 80 }, { name: 'Murray', age: 68 }, ...]
-      return notes
-    }
-  
+import connectDB from '../../config/db';
+import Order from '../../models/Order';
 
-
+async function getNotes() {
+  await connectDB();
+  const notes = await Order.find({ partner_id: '44' }).select([
+    'name', 'state', 'client_order_ref', 'origin', 'reference', 'signed_by',
+    'invoice_status', 'validity_date', 'note', 'currency_rate', 'amount_untaxed',
+    'amount_tax', 'amount_total', 'amount_to_invoice', 'locked', 'require_signature',
+    'require_payment', 'create_date', 'commitment_date', 'date_order', 'signed_on',
+    'write_date', 'prepayment_percent', 'sale_order_template_id'
+  ]);
+  return notes;
+}
 
 let notes = getNotes()
 let  notasPartidas: {
@@ -86,8 +88,6 @@ console.log(hoy,note.name,fechaCreacion,
 )
 
 )
-
-*/
 
 export default async function Productos() {
   // Render the main container
