@@ -9,6 +9,14 @@ import Table from 'react-bootstrap/Table';
 import axios from 'axios';
 import UpdateandDeleteControls from './updateanddelete';
 import FacturaGlobalMaker from '../libComponents/facturaGlobalMaker';
+import Container from 'react-bootstrap/Container';
+import Row from 'react-bootstrap/Row';
+import Col from 'react-bootstrap/Col';
+import Receipt58mm from './receipt58mm';
+import dias from './data.json'
+import postRequest from "../libComponents/postRequest";
+import { toast } from 'react-toastify';
+const  orderID = require('order-id')('key');
 interface FormCreatorProps {
   elements: {
     name: string;
@@ -21,32 +29,130 @@ interface FormCreatorProps {
 
 function FormPos(props:FormCreatorProps) {
   const [show, setShow] = useState(false);
+  const [showChangeModal, setShowChangeModal] = useState(false);
   const [ticketContent, setTicketContent] = useState('');
-  const [formData1, setFormData1] = useState(new FormData());
-  const handleTicketContent = () => setTicketContent(ticket)
+  const [amountPaid, setAmountPaid] = useState<number>(0);
+  const [total, setTotal] = useState<number>(0);
+  const [notasPartidas, setNotasPartidas] = useState<any[]>([]);
+  const [empleado, setEmpleado] = useState<number>(0);
+  const [empleados, setEmpleados] = useState<Array<{ id: number; nombreCompleto: string }>>([]);
+
+  // Fetch employees when component mounts
+  useEffect(() => {
+    const fetchEmpleados = async () => {
+      try {
+        const response = await fetch('/api/empleados');
+        if (response.ok) {
+          const data = await response.json();
+          setEmpleados(data);
+          if (data.length > 0) {
+            setEmpleado(data[0].id); // Set first employee as default
+          }
+        }
+      } catch (error) {
+        console.error('Error fetching empleados:', error);
+        toast.error('Error al cargar empleados');
+      }
+    };
+    fetchEmpleados();
+  }, []);
+
+  //const handleTicketContent = () => setTicketContent(ticket)
   const handleClose = () => setShow(false);
   const handleShow = () => setShow(true);
-  const [tableTicketElements, setTableTicketElements] = useState<any[]>([]);
+  const handleChangeModalClose = async () => {
+    const change = amountPaid - total;
+    setShowChangeModal(false);
+    
+    try {
+      // Guardar el recibo en la base de datos
+      const receiptData = {
+        id: folio,
+        amount: total,
+        date: new Date().toISOString(),
+        status: 'completed',
+        empleadoId: empleado,
+        items: notasPartidas.map(item => ({
+          quantity: item.Cantidad,
+          description: item.Descripcion,
+          unitPrice: Number(item.ValorUnitario),
+          total: Number(item.ImporteRealConImp)
+        }))
+      };
 
+      await fetch('/api/receipts', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify(receiptData),
+      });
+
+      // Generar el recibo y imprimir
+      //ReciboMaker(notasPartidas, folio, empleado);
+      print();
+
+      // Limpiar el estado después de la venta
+      setNotasPartidas([]);
+      setAmountPaid(0);
+      setTotal(0);
+    } catch (error) {
+      console.error('Error saving receipt:', error);
+      toast.error('Error al guardar el recibo');
+    }
+  };
+  const handleChangeModalShow = () => {
+    // Calculate total from notasPartidas
+    const newTotal = notasPartidas.reduce((sum, item) => sum + Number(item.ImporteRealConImp), 0);
+    setTotal(newTotal);
+    setShowChangeModal(true);
+  };
+
+  const calculateChange = () => {
+    return amountPaid - total;
+  };
+let  folio = orderID.generate(); 
   let ticket:any;
-  let elements=props.elements
-  let formElements:any[]=[]
-  var optionelements:any[]=[]
-
-   elements.map((element)=>{
-  formElements.push(
-    <Form.Group className="mb-3" controlId={element.name+element.id} key={element.id+"key1"}>
-    <Form.Label key={element.id+"key2"}>{element.placeholder}</Form.Label>
-    <Form.Control  key={element.id+"key3"}type={element.type} name={element.name} placeholder={element.placeholder} />
+  let elements = props.elements
   
-  </Form.Group>
-  )
-  return formElements
-  })
+  // Create form elements array
+  const formElements = [
+    // Add employee select field at the beginning of the form
+    <Form.Group className="mb-3" controlId="empleado-select" key="empleado-select">
+      <Form.Label>Empleado</Form.Label>
+      <Form.Select 
+        required 
+        value={empleado}
+        onChange={(e) => setEmpleado(Number(e.target.value))}
+      >
+        {empleados.map(emp => (
+          <option key={emp.id} value={emp.id}>
+            {emp.nombreCompleto}
+          </option>
+        ))}
+      </Form.Select>
+    </Form.Group>,
+    
+    // Map through the elements array to create form groups
+    ...elements.map((element) => (
+      <Form.Group 
+        className="mb-3" 
+        controlId={`${element.name}-${element.id}`} 
+        key={`form-group-${element.id}`}
+      >
+        <Form.Label>{element.placeholder}</Form.Label>
+        <Form.Control
+          required
+          type={element.type}
+          name={element.name}
+          placeholder={element.placeholder}
+        />
+      </Form.Group>
+    ))
+  ];
 
-  let notasPartidas:any[]=[]
-  function addPartida(pu:any,cantidad) {
-     
+  function addPartida(pu:any,cantidad:any,Descripcion:any) {
+  
     let importeConImp=pu*cantidad
     let iva = Number(importeConImp) * 0.16
     let impSinImp=Number(importeConImp-iva)
@@ -55,11 +161,12 @@ function FormPos(props:FormCreatorProps) {
       let partida= {
         ClaveProdServ: "01010101",
         ClaveUnidad: "ACT",
+        NoIdentificacion: ""+folio+"",
         Unidad: "Actividad",
         Cantidad: cantidad,
-        Descripcion: "Venta",
-        ValorUnitario: ""+puSinImp+"",
-        Importe: ""+impSinImp+"",
+        Descripcion:Descripcion ,
+        ValorUnitario: ""+pu+"",
+        Importe: ""+impSinImp.toFixed(2)+"",
         ImporteRealConImp:"" +Number(importeConImp) +"",
         ObjetoImp:"02",
         Impuestos :{
@@ -69,7 +176,7 @@ function FormPos(props:FormCreatorProps) {
           Impuesto: "002",
           TipoFactor: "Tasa",
           TasaOCuota: "0.160000",
-          Importe: ""+iva+"",
+          Importe: ""+iva.toFixed(2)+"",
           }
           ]
       
@@ -78,544 +185,294 @@ function FormPos(props:FormCreatorProps) {
         }
        
       notasPartidas.push(partida)
+ 
+      setNotasPartidas([...notasPartidas])
 
     }
-    function print(ticket) {
-      var printContents = document.getElementById("ticket").innerHTML;
-      var originalContents = document.body.innerHTML;
-      document.body.innerHTML = printContents;
-    
-      window.print();
-    
-      document.body.innerHTML = originalContents;
+    function print() {
+      const receiptElement = document.getElementById('ticket-div');
+         // Generate folio directly in the print function
+      if (receiptElement) {
+        const printWindow = window.open('', '_blank');
+        if (printWindow) {
+          printWindow.document.write(`
+            <html>
+              <head>
+                <title>Ticket de Venta</title>
+                <style>
+                  @page {
+                    size: 80mm auto;
+                    margin: 0;
+                  }
+                  body {
+                    font-family: 'Arial', sans-serif;
+                    width: 80mm;
+                    margin: 0;
+                    padding: 10px;
+                  }
+                  #print-content {
+                    text-align: center;
+                  }
+                  .header {
+                    font-size: 14px;
+                    margin-bottom: 10px;
+                  }
+                  .company-name {
+                    font-size: 16px;
+                    font-weight: bold;
+                    margin-bottom: 5px;
+                  }
+                  .info {
+                    font-size: 12px;
+                    margin-bottom: 5px;
+                  }
+                  .divider {
+                    border-top: 1px dashed #000;
+                    margin: 10px 0;
+                  }
+                  .item {
+                    text-align: left;
+                    font-size: 12px;
+                    margin: 5px 0;
+                  }
+                  .totals {
+                    text-align: right;
+                    font-size: 12px;
+                    margin-top: 10px;
+                  }
+                  .footer {
+                    font-size: 12px;
+                    text-align: center;
+                    margin-top: 10px;
+                  }
+                  @media print {
+                    .no-print {
+                      display: none;
+                    }
+                  }
+                </style>
+              </head>
+              <body>
+                <div id="print-content">
+                  <div class="company-name">Bici-vic</div>
+                  <div class="header">
+                    <div class="info">Sucursal: 5 Matamoros</div>
+                    <div class="info">RFC:FOZA8801257C2</div>
+                    <div class="info">Dirección: Mariano Matamoros #1242 Zona Centro</div>
+                    <div class="info">Tel: (834) 2027887</div>
+                    <div class="info">C.P: 87000 Ciudad Victoria,Tamps.</div>
+                    <div class="info">Regimen Simplificado de Confianza</div>
+                    </div>
+                  <div class="divider"></div>
+                  <div class="info">Ticket #: ${folio}</div>
+                  <div class="info">Fecha: ${new Date().toLocaleString()}</div>
+                  <div class="divider"></div>
+                  ${receiptElement.innerHTML}
+                  <div class="divider"></div>
+                  <div class="footer">
+                    ¡Gracias por su compra!
+                    <br>
+                    Conserve su ticket
+                    Expedido en: Ciudad Victoria, Tamps.
+                  </div>
+                </div>
+                <script>
+                  window.onload = function() {
+                    window.print();
+                    window.onafterprint = function() {
+                      window.close();
+                    };
+                  };
+                </script>
+              </body>
+            </html>
+          `);
+          printWindow.document.close();
+        }
+      }
     }
-  function arrayToPartidas(array){
+  function arrayToPartidas(array:[]){
     let impsConvertidos=[]
     array.map((imp:any)=>{
-      addPartida(imp,1)
+      addPartida(imp,1,"Venta")
     })
     impsConvertidos=notasPartidas
+    
 
     return impsConvertidos
   }
-let dos_dic=[50,50,60,55,35,50,145,70,15,48,298,340,36,36,10,15,98,195,80,20,10,18,220,15,10,10,110,30,90,7,12,225,2629,75,50,54]
-//let uno_dic_partidas=arrayToPartidas(uno_dic)  
-let tres_dic=[180,170,50,310,135,45,160,180,12,60,180,125,175,80,45]
-//let fact=ReciboMaker(uno_dic_partidas)
 
 
-
-var dia19=[50,75,55,60,15,15,20,15,205,45,45,60,60,50,15,75,25,250]
-var dia20=[86,6,120,1500,155,195,195,70]
-let veintiuno_nov=[65,675,150,60,150,450,45,35,200,40,75,10,10,150,5,40,10,40,7,40,20,15,21,18,45,150,410,40,85]
-let veintidos_nov=[167,60,75,50,10,15,36,15,205,80,45,34,92,45,45,75,50,50,10,95,20,80]
-let veintitres_nov=[139,120,12,30,705,220,70,30,40,377,75]
-let nov25=[240,80,45,35,25,120,30,70,120,75,40,50,340,900,15,385,120,135,65,180,120,300,260]
-let nov26=[85,400,160,130,540,92,160,120,20,12,140,120,540,465,80,15,160,175,8,45,40,45,795,70,50,52,600,36,36,40,40,500]
-let nov27=[360,172,65,95,30,45,90,45,219,75,95,8,135,25,10,30,30,540,540,100] 
-let nov28=[60,85,330,330,220,105,105,45,45,105,15,45,500,75,80,40,130]
-let nov29=[220,60,45,45,50,95,7,70,21,55,30,75,30,10,80]
-let nov30=[340,50,10,7,160,50,35,36,40,55,75,30,110,80,20,80,85]
-let nov_pedidos=[1500,600,1000,2000,180,430,250,1450]
-let nov=dia19.concat(dia20,veintiuno_nov,veintidos_nov,veintitres_nov,nov25,nov26,nov27,nov28,nov29,nov30,nov_pedidos)
-
-
-let conv=arrayToPartidas(nov)
-
-//let fact=ReciboMaker(conv)
-
-let imps_unoaldieciocho=[
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "75.6",
-        "Importe": "75.6",
-        "ImporteRealConImp": "90",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "90",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "14.4"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "159.6",
-        "Importe": "159.6",
-        "ImporteRealConImp": "190",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "190",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "30.40"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "60.48",
-        "Importe": "60.48",
-        "ImporteRealConImp": "72",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "72",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "11.52"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "105.84",
-        "Importe": "105.84",
-        "ImporteRealConImp": "126",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "126",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "20.16"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "204.12",
-        "Importe": "204.12",
-        "ImporteRealConImp": "243",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "243",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "38.88"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "25.2",
-        "Importe": "25.2",
-        "ImporteRealConImp": "30",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "30",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "4.8"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "315",
-        "Importe": "315",
-        "ImporteRealConImp": "375",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "375",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "60"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "134.4",
-        "Importe": "134.4",
-        "ImporteRealConImp": "160",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "160",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "25.6"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "92.4",
-        "Importe": "92.4",
-        "ImporteRealConImp": "110",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "110",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "17.6"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "117.6",
-        "Importe": "117.6",
-        "ImporteRealConImp": "140",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "140",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "22.40"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "84",
-        "Importe": "84",
-        "ImporteRealConImp": "100",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "100",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "16"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "252",
-        "Importe": "252",
-        "ImporteRealConImp": "300",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "300",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "48"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "25.2",
-        "Importe": "25.2",
-        "ImporteRealConImp": "30",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "30",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "4.8"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "151.2",
-        "Importe": "151.2",
-        "ImporteRealConImp": "180",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "180",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "28.8"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "50.4",
-        "Importe": "50.4",
-        "ImporteRealConImp": "60",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "60",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "9.6"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "185.64",
-        "Importe": "185.64",
-        "ImporteRealConImp": "221",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "221",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "35.36"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "283.08",
-        "Importe": "283.08",
-        "ImporteRealConImp": "337",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "337",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "53.92"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "92.4",
-        "Importe": "92.4",
-        "ImporteRealConImp": "110",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "110",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "17.6"
-                }
-            ]
-        }
-    },
-    {
-        "ClaveProdServ": "01010101",
-        "ClaveUnidad": "ACT",
-        "Unidad": "Actividad",
-        "Cantidad": 1,
-        "Descripcion": "Venta",
-        "ValorUnitario": "144.48",
-        "Importe": "144.48",
-        "ImporteRealConImp": "172",
-        "ObjetoImp": "02",
-        "Impuestos": {
-            "Traslados": [
-                {
-                    "Base": "172",
-                    "Impuesto": "002",
-                    "TipoFactor": "Tasa",
-                    "TasaOCuota": "0.160000",
-                    "Importe": "27.52"
-                }
-            ]
-        }
-    }
-]
-   
-let nov_notas=imps_unoaldieciocho.concat(conv)
- 
-
-
-
-  function onSubmitForRecibo(FormData: FormData) {
-    ticket=ReciboMaker(nov_notas) 
-    handleTicketContent()
-    handleShow()
-  }
   function onSubmit(formData: FormData) {
     let entries = Object.fromEntries(formData.entries()); 
     console.log(entries);
-    //addPartida(entries.importe,entries.Cantidad)
-    
-
+    addPartida(entries.importe,entries.Cantidad,entries.Descripcion)
+    //send to state
+    setNotasPartidas([...notasPartidas]);
   }
  
 
    
+  const handleCashClosingShow = () => {
+    setShow(true);
+  };
+
+  const handleCashClosingClose = () => {
+    setShow(false);
+  };
+
   return (
     <>
-    <Form action={onSubmit}>
+    <Container>
+      <Row>
+        <Col>
+        <div style={{
+          backgroundColor: 'white',
+          padding: '25px',
+          borderRadius: '12px',
+          boxShadow: '0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06)',
+          margin: '20px 0',
+          border: '1px solid #f0f0f0'
+        }}>
+        <h3 style={{ 
+          color: '#2d3748',
+          marginBottom: '20px',
+          fontSize: '1.5rem',
+          fontWeight: '600'
+        }}>Ingresa aquí los productos que vendiste:</h3>
+         <Form action={onSubmit}>
       {formElements}
-      <Button variant="primary" type="submit" >
+      <Button variant="primary" type="submit" style={{
+        marginTop: '15px',
+        padding: '10px 20px',
+        borderRadius: '8px',
+        width: '100%'
+      }}>
         Agregar
       </Button>
     </Form>
-   <hr/>
+        </div>
+        </Col>  
+        <Col>
+        <div style={{ 
+          backgroundColor: 'white',
+          padding: '25px',
+          borderRadius: '12px',
+          boxShadow: '0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06)',
+          margin: '20px 0',
+          border: '1px solid #f0f0f0'
+        }} >
+        <h3 style={{ 
+          color: '#2d3748',
+          marginBottom: '20px',
+          fontSize: '1.5rem',
+          fontWeight: '600'
+        }}>Ticket</h3>
 
-    <Form action={onSubmitForRecibo}>   
-    <Table striped>
-      <thead>
-        <tr>
-          <th>Cantidad</th>
-          <th>Desc.</th>
-          <th>P.U.</th>
-          <th>Importe</th>
-        </tr>
-      </thead>
-      <tbody>
-        {/* { Array.isArray(tableTicketElements.map((element:any) => (
-          <tr key={element._id}>
-            <td>{element.Cantidad}</td>
-            <td>{element.Descripcion}</td>
-            <td>{(element.ImporteRealConImp/element.Cantidad)}</td>
-            <td>{element.ImporteRealConImp}</td>
-          </tr>
-        )))} */}
-      </tbody>
-    </Table>
-    <Button variant="success" type="submit" >
+        <div style={{ overflowX: 'auto' }} id='ticket-div'>
+          <Table hover style={{ 
+            fontSize: '0.9em',
+            backgroundColor: 'white',
+            borderRadius: '8px',
+            overflow: 'hidden',
+            marginBottom: '20px'
+          }}>
+            <thead style={{
+              backgroundColor: '#f8f9fa',
+              borderBottom: '2px solid #dee2e6'
+            }}>
+              <tr>
+                <th style={{ padding: '12px 15px' }}>Cantidad</th>
+                <th style={{ padding: '12px 15px' }}>Descripción</th>
+                <th style={{ padding: '12px 15px' }}>P.U.</th>
+                <th style={{ padding: '12px 15px' }}>Importe</th>
+              </tr>
+            </thead>
+            <tbody>
+              {notasPartidas.map((item: any, index: number) => (
+                <tr key={index} style={{ 
+                  borderBottom: '1px solid #dee2e6',
+                  transition: 'background-color 0.2s'
+                }}>
+                  <td style={{ padding: '12px 15px' }}>{item.Cantidad}</td>
+                  <td style={{ padding: '12px 15px' }}>{item.Descripcion}</td>
+                  <td style={{ padding: '12px 15px' }}>${Number(item.ValorUnitario).toFixed(2)}</td>
+                  <td style={{ padding: '12px 15px' }}>${Number(item.ImporteRealConImp).toFixed(2)}</td>
+                </tr>
+              ))}
+              <tr style={{ 
+                borderTop: '2px solid #dee2e6',
+                backgroundColor: '#f8f9fa',
+                fontWeight: 'bold'
+              }}>
+                <td colSpan={3} style={{ padding: '12px 15px', textAlign: 'right' }}>Total:</td>
+                <td style={{ padding: '12px 15px' }}>${notasPartidas.reduce((sum, item) => sum + Number(item.ImporteRealConImp), 0).toFixed(2)}</td>
+              </tr>
+            </tbody>
+          </Table>
+        </div>
+
+        <Form >   
+    <div style={{ borderTop: '2px dashed #ccc', marginTop: '15px', paddingTop: '15px', textAlign: 'center' }}>
+    <Button variant="success" type="submit" style={{ width: '100%', marginBottom: '10px' }} onClick={(e) => {
+      e.preventDefault();
+      handleChangeModalShow();
+    }}>
     Crear Ticket
     </Button>
+    </div>
     </Form>
+   
+        </div>
+        </Col>
+      </Row>
+    </Container>
+       
+      
 
-
-      <Modal show={show} onHide={handleClose}>
+      <Modal show={showChangeModal} onHide={handleChangeModalClose}>
         <Modal.Header closeButton>
-          <Modal.Title>Ticket</Modal.Title>
+          <Modal.Title>Pago</Modal.Title>
         </Modal.Header>
-        <Modal.Body>{ticketContent}</Modal.Body>
+        <Modal.Body>
+          <Form>
+            <Form.Group className="mb-3">
+              <Form.Label>Total a pagar</Form.Label>
+              <Form.Control
+                type="number"
+                value={total}
+                readOnly
+              />
+            </Form.Group>
+            <Form.Group className="mb-3">
+              <Form.Label>Cantidad pagada</Form.Label>
+              <Form.Control
+                type="number"
+                value={amountPaid}
+                onChange={(e) => setAmountPaid(Number(e.target.value))}
+                autoFocus
+              />
+            </Form.Group>
+            <Form.Group className="mb-3">
+              <Form.Label>Cambio</Form.Label>
+              <Form.Control
+                type="number"
+                value={amountPaid - total}
+                readOnly
+              />
+            </Form.Group>
+          </Form>
+        </Modal.Body>
         <Modal.Footer>
-          <Button variant="secondary" onClick={handleClose}>
+          <Button variant="secondary" onClick={() => setShowChangeModal(false)}>
             Cerrar
           </Button>
-          <Button variant="success" onClick={handleClose}>
+          <Button variant="primary" onClick={handleChangeModalClose}>
             Imprimir
           </Button>
         </Modal.Footer>
       </Modal>
+
     </>
   );
 }
